Fix 30 Minute Login Failure
## Problem:
Problem arose due to changing HttpClient to static. Because getting the
login parameters and Posting the login parameters to get the "ticket"
hits the same URL, and we're using the same HttpClient instance,
subsequent requests contain the auth "cookie". If you do a GET to that
url with a valid AuthCookie, it updates the "ticket" needed to get the
OAuthToken.

## Solution
Allow login request 2 of 3 to accept null parameters, and do a GET
instead of a POST if the parameter is null.

STATUS: Awaiting final testing